requirements.txt
fastapi
uvicorn
python-jose
pydantic[dotenv]
pydantic
passlib
python-multipart
bcrypt
requests
cryptography
tqdm